requirements.txt
MetaTrader5~=5.0.4200
pandas~=2.2.0
plotly
ta
Pyarrow
IPython~=8.22.1
pytz~=2024.1
python-dateutil~=2.8.2
matplotlib~=3.8.3
mplfinance~=0.12.10b0
requests